requirements.txt
Flask
tensorflow
keras
numpy
Pillow
psycopg2
Requests
gunicorn